# specs/001-pointsite-builder/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PointSite Builder API
  version: 1.0.0
  description: Private API using GitHub collaborator sessions. All mutation responses include a request ID.
servers:
  - url: /api
security:
  - githubSession: []
paths:
...
  /publish/staging:
    post:
      operationId: publishStaging
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [draftId, expectedBaseSha]
              properties:
                draftId: { type: string, format: uuid }
                expectedBaseSha: { type: string, pattern: '^[a-f0-9]{40}$' }
      responses:
        '201':
          description: Staging candidate commit and durable publish job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StagingPublishResult' }
        '409': { $ref: '#/components/responses/Error' }
...
  /publish/production:
    post:
      operationId: prepareProductionPullRequest
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '403':
          description: Always disabled until production-preparation configuration and approval exist
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /publish/jobs/{jobId}:
    get:
      operationId: getJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Exact publish job and evidence
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublishJob' }
  /publish/jobs/{jobId}/verification:
    post:
      operationId: refreshStagingVerification
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Publish job with exact GitHub check evidence
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublishJob' }
...
  /approvals:
    post:
      operationId: recordApproval
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [publishJobId, expectedTuple, decision]
              properties:
                publishJobId: { type: string, minLength: 1, maxLength: 100 }
                expectedTuple: { $ref: '#/components/schemas/CandidateTuple' }
                decision: { type: string, enum: [approved, rejected, revoked] }
                note: { type: string, maxLength: 500 }
      responses:
        '201': { description: Recorded approval }
        '409': { $ref: '#/components/responses/Error' }
  /approvals/eligibility:
    post:
      operationId: checkStagingEligibility
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [tuple, observedProductionBaseSha]
              properties:
                tuple: { $ref: '#/components/schemas/CandidateTuple' }
                observedProductionBaseSha: { type: string, pattern: '^[a-f0-9]{40}$' }
      responses:
        '200': { description: Current exact-candidate eligibility }
  /admin/roles:
    get:
      operationId: listRoles
      responses:
        '200': { description: Active and disabled role mappings }
    put:
      operationId: upsertRole
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [githubLogin, role, active]
              properties:
                githubLogin: { type: string, pattern: '^[A-Za-z0-9-]{1,39}$' }
                role: { type: string, enum: [viewer, editor, publisher, administrator] }
                active: { type: boolean }
      responses:
        '200': { description: Role mapping }
  /admin/audit:
    get:
      operationId: listAudit
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200': { description: Paginated audit events }
...
  /admin/retention/apply:
    post:
      operationId: applyRetention
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [exportChecksum]
              properties:
                exportChecksum: { type: string, pattern: '^[a-f0-9]{64}$' }
      responses:
        '200': { description: 'Applied exact, previously exported retention plan' }
        '409': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    githubSession:
      type: apiKey
      in: cookie
      name: __Secure-pointsite_builder_session
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 100 }
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    PublishJob:
      type: object
      required: [id, environment, status, candidateChecksum, requestedBy, requestedAt]
      properties:
        id: { type: string, format: uuid }
        environment: { type: string, enum: [staging, production-pr, production-merge, rollback] }
        status: { type: string, enum: [queued, running, succeeded, failed, cancelled] }
        candidateChecksum: { type: string }
        resultSha: { type: string }
        externalUrl: { type: string, format: uri }
        requestedBy: { type: string }
        requestedAt: { type: string, format: date-time }
        evidence: { type: object }
...
    CandidateTuple:
      type: object
      additionalProperties: false
      required:
        [
          siteId,
          revisionId,
          revisionChecksum,
          schemaVersion,
          rendererVersion,
          candidateChecksum,
          stagingBaseSha,
          stagingCommitSha,
          productionBaseSha,
        ]
      properties:
        siteId: { type: string, const: pointsite }
        revisionId: { type: string, format: uuid }
        revisionChecksum: { type: string, pattern: '^[a-f0-9]{64}$' }
        schemaVersion: { type: integer, minimum: 1 }
        rendererVersion: { type: string }
        candidateChecksum: { type: string, pattern: '^[a-f0-9]{64}$' }
        stagingBaseSha: { type: string, pattern: '^[a-f0-9]{40}$' }
        stagingCommitSha: { type: string, pattern: '^[a-f0-9]{40}$' }
        productionBaseSha: { type: string, pattern: '^[a-f0-9]{40}$' }
    StagingPublishResult:
      type: object
      additionalProperties: false
      required:
        [
          environment,
          siteId,
          revisionId,
          revisionChecksum,
          schemaVersion,
          rendererVersion,
          candidateChecksum,
          baseSha,
          stagingBaseSha,
          commitSha,
          url,
          requestedBy,
          status,
        ]
      properties:
        environment: { type: string, const: staging }
        siteId: { type: string, const: pointsite }
        revisionId: { type: string, format: uuid }
        revisionChecksum: { type: string, pattern: '^[a-f0-9]{64}$' }
        schemaVersion: { type: integer, minimum: 1 }
        rendererVersion: { type: string }
        candidateChecksum: { type: string, pattern: '^[a-f0-9]{64}$' }
        baseSha: { type: string, pattern: '^[a-f0-9]{40}$' }
        stagingBaseSha: { type: string, pattern: '^[a-f0-9]{40}$' }
        commitSha: { type: string, pattern: '^[a-f0-9]{40}$' }
        url: { type: string, format: uri }
        jobId: { type: string }
        requestedBy: { type: string }
        status: { type: string, const: succeeded }
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            required: [path, message]
            properties:
              path: { type: string }
              message: { type: string }